+++ src/__typetests__/__fixtures__/dummy.d.ts
