+++ src/__typetests__/__fixtures__/pathString.ts
