+++ src/__typetests__/__fixtures__/traversable.ts
