+++ src/__typetests__/types/package.json
@@ -1,0 +1,6 @@
+{
+  "types": "./__fixtures__/dummy.d.ts",
+  "tsd": {
+    "directory": "."
+  }
+}
